+++ projects/usermusl/src/thread/pthread_getname_np.c
@@ -13,7 +13,7 @@ int pthread_getname_np(pthread_t thread, char *name, size_t len)
 	if (len < 16) return ERANGE;
 
 	if (thread == pthread_self())
-		return prctl(PR_GET_NAME, (unsigned long)name, 0UL, 0UL, 0UL) ? errno : 0;
+		return prctl(PR_GET_NAME, name, 0UL, 0UL, 0UL) ? errno : 0;
 
 	snprintf(f, sizeof f, "/proc/self/task/%d/comm", thread->tid);
 	pthread_setcancelstate(PTHREAD_CANCEL_DISABLE, &cs);
